refactor: extract booths, jobInvitations, and resume routes
- Extract booth endpoints (create, logo upload/url, visitor tracking, ratings) to routes/booths.js
- Extract job invitation endpoints (send, received, sent, status, stats, details, by-id) to routes/jobInvitations.js
- Extract resume endpoints (upload, url, parse, tailor, tailored CRUD) to routes/resume.js
- Remove all extracted handlers from server.js; mount new routers via app.use("/api", ...)
- Fix inner require path for resumeParser in upload-resume handler
